# cmd/api/users.go
package main

import (
	"Go-Microservice/internal/repo"
	"context"
	"net/http"
	"strconv"

	"github.com/go-chi/chi/v5"
)

type userKey string

const userCtx userKey = "user"

type FollowerUser struct {
	UserId int64 `json:"userId"`
}

type FollowedUser struct {
	UserID     int64 `json:"userId"`
	FollowerId int64 `json:"followerId"`
}

// GetUser retrieves a specific user by their ID
//
//	@Summary		Get user by ID
//	@Description	Retrieve detailed information about a specific user including their profile data
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			userID	path		int64				true	"User ID to retrieve"
//	@Success		200		{object}	repo.User			"User details retrieved successfully"
//	@Failure		400		{object}	map[string]string	"Invalid user ID format"
//	@Failure		404		{object}	map[string]string	"User not found"
//	@Failure		500		{object}	map[string]string	"Internal server error"
//	@Router			/v1/users/{userID} [get]
func (app *application) getUserHandler(w http.ResponseWriter, r *http.Request) {
	user := getUserFromContext(r)
	if err := app.jsonResponse(w, http.StatusOK, user); err != nil {
		app.internalServerError(w, r, err)
	}
}

// FollowUser allows the current user to follow another user
//
//	@Summary		Follow a user
//	@Description	Create a following relationship between the authenticated user and the target user
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			userID	path		int64				true	"ID of the user to follow"
//	@Success		200		{object}	map[string]string	"Successfully followed user"
//	@Success		201		{object}	map[string]string	"Following relationship created"
//	@Failure		400		{object}	map[string]string	"Invalid user ID or cannot follow yourself"
//	@Failure		404		{object}	map[string]string	"User not found"
//	@Failure		409		{object}	map[string]string	"Already following this user"
//	@Failure		500		{object}	map[string]string	"Internal server error"
//	@Security		BasicAuth
//	@Router			/v1/users/{userID}/follow [put]
func (app *application) followUserHandler(w http.ResponseWriter, r *http.Request) {
	followerUser := getUserFromContext(r)

	var payload FollowedUser
	if err := readJSON(w, r, &payload); err != nil {
		app.badRequestResponse(w, r, err)
		return
	}
	ctx := r.Context()
	if err := app.repo.Followers.Follow(ctx, followerUser.ID, payload.UserID); err != nil {
		switch err {
		case repo.ErrConflict:
			app.conflictResponse(w, r, err)
			return
		default:
			app.internalServerError(w, r, err)
			return
		}
	}

	if err := app.jsonResponse(w, http.StatusNoContent, nil); err != nil {
		app.internalServerError(w, r, err)
	}
}

// UnfollowUser allows the current user to unfollow another user
//
//	@Summary		Unfollow a user
//	@Description	Remove the following relationship between the authenticated user and the target user
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			userID	path		int64				true	"ID of the user to unfollow"
//	@Success		200		{object}	map[string]string	"Successfully unfollowed user"
//	@Success		204		{object}	map[string]string	"Following relationship removed"
//	@Failure		400		{object}	map[string]string	"Invalid user ID or cannot unfollow yourself"
//	@Failure		404		{object}	map[string]string	"User not found or not following"
//	@Failure		500		{object}	map[string]string	"Internal server error"
//	@Security		BasicAuth
//	@Router			/v1/users/{userID}/unfollow [put]
func (app *application) unfollowUserHandler(w http.ResponseWriter, r *http.Request) {
	unFollowingUser := getUserFromContext(r)

	var payload FollowedUser
	if err := readJSON(w, r, &payload); err != nil {
		app.badRequestResponse(w, r, err)
		return
	}
	ctx := r.Context()

	if err := app.repo.Followers.Unfollow(ctx, unFollowingUser.ID, payload.UserID); err != nil {
		app.internalServerError(w, r, err)
		return
	}

	if err := app.jsonResponse(w, http.StatusNoContent, nil); err != nil {
		app.internalServerError(w, r, err)
	}
}

func (app *application) usersContextMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		userID, err := strconv.ParseInt(chi.URLParam(r, "userID"), 10, 64)
		if err != nil {
			app.badRequestResponse(w, r, err)
			return
		}

		user, err := app.repo.Users.GetByID(r.Context(), userID)
		if err != nil {
			switch err {
			case repo.ErrPostNotFound:
				app.notFoundResponse(w, r, err)
				return
			default:
				app.internalServerError(w, r, err)
				return
			}
		}

		ctx := context.WithValue(r.Context(), userCtx, user)
		next.ServeHTTP(w, r.WithContext(ctx))
	})
}

func getUserFromContext(r *http.Request) *repo.User {
	user, _ := r.Context().Value(userCtx).(*repo.User)
	return user
}